+++ Intersectii/vector2D.cpp
@@ -7,7 +7,7 @@
 
 #include "vector2D.h"
 
-float Vector2D::CrossProduct(const Vector2D & other) const
+double Vector2D::CrossProduct(const Vector2D & other) const
 {
 	return (this->XVect * other.YVect - this->YVect * other.XVect);
 }
